+++ src/Facade/Cpu/OSXFacade.php
@@ -8,31 +8,55 @@ use Innmind\Server\Status\{
     Server\Cpu\Percentage,
     Server\Cpu\Cores,
 };
+use Innmind\Server\Control\Server\{
+    Processes,
+    Command,
+};
 use Innmind\Immutable\{
     Str,
     Maybe,
 };
-use Symfony\Component\Process\Process;
 
 /**
  * @internal
  */
 final class OSXFacade
 {
+    private Processes $processes;
+
+    public function __construct(Processes $processes)
+    {
+        $this->processes = $processes;
+    }
+
     /**
      * @return Maybe<Cpu>
      */
     public function __invoke(): Maybe
     {
-        $process = Process::fromShellCommandline('top -l 1 -s 0 | grep \'CPU usage\'');
-        $process->run();
+        return $this
+            ->processes
+            ->execute(
+                Command::foreground('top')
+                    ->withShortOption('l', '1')
+                    ->withShortOption('s', '0')
+                    ->pipe(
+                        Command::foreground('grep')
+                            ->withArgument('CPU usage'),
+                    ),
+            )
+            ->wait()
+            ->maybe()
+            ->map(static fn($success) => $success->output()->toString())
+            ->flatMap($this->parse(...));
+    }
 
-        if (!$process->isSuccessful()) {
-            /** @var Maybe<Cpu> */
-            return Maybe::nothing();
-        }
-
-        $percentages = Str::of($process->getOutput())
+    /**
+     * @return Maybe<Cpu>
+     */
+    private function parse(string $output): Maybe
+    {
+        $percentages = Str::of($output)
             ->trim()
             ->capture(
                 '~^CPU usage: (?P<user>\d+\.?\d*)% user, (?P<sys>\d+\.?\d*)% sys, (?P<idle>\d+\.?\d*)% idle$~',
@@ -40,13 +64,23 @@ final class OSXFacade
             ->map(static fn($_, $percentage) => $percentage->toString())
             ->map(static fn($_, $percentage) => (float) $percentage);
 
-        $process = Process::fromShellCommandline('sysctl -a | grep \'hw.ncpu:\'');
-        $process->run();
-
-        $cores = Str::of($process->getOutput())
-            ->trim()
-            ->capture('~^hw.ncpu: (?P<cores>\d+)$~')
-            ->get('cores')
+        $cores = $this
+            ->processes
+            ->execute(
+                Command::foreground('sysctl')
+                    ->withShortOption('a')
+                    ->pipe(
+                        Command::foreground('grep')
+                            ->withArgument('hw.ncpu:'),
+                    ),
+            )
+            ->wait()
+            ->maybe()
+            ->map(static fn($success) => $success->output()->toString())
+            ->map(Str::of(...))
+            ->map(static fn($output) => $output->trim())
+            ->map(static fn($output) => $output->capture('~^hw.ncpu: (?P<cores>\d+)$~'))
+            ->flatMap(static fn($output) => $output->get('cores'))
             ->map(static fn($cores) => $cores->toString())
             ->map(static fn($cores) => (int) $cores)
             ->otherwise(static fn() => Maybe::just(1));